feat(collect): scheduled daily collection, and the fixes it exposed
Adds .github/workflows/collect.yml (daily 06:00 UTC, manual dispatch, a
concurrency guard so runs cannot overlap) and docs/runbooks/collection.md
covering the cadence trade-off, server cron/systemd alternatives, and triage.

Setting the job up surfaced three defects that only appear under repetition:

* The stay grid decayed. --bootstrap only fired for hotels with zero
  observations and pinned check-in dates once, while planCollection only
  refreshes stays with check_in >= CURRENT_DATE. The tracked set would have
  aged out one day at a time and emptied after ~90 days, with every run still
  exiting 0. The grid is now topped up on every run from lead times relative
  to today (findMissingGridStays).

* planCollection was broken and had never been executed. `CURRENT_DATE + $1`
  leaves Postgres unable to choose between date+integer and date+interval:
  "operator is not unique". Every prior run used the cold-start path, so this
  would have failed on the SECOND day of scheduled collection. It is pure SQL,
  so only an integration test can catch it — two now do.

* Stays that yield nothing were re-requested forever. Because the grid is
  rebuilt from lead-time offsets, a stay that is sold out or that the API
  refuses with a 500 still looks "missing" next run. Measured on the live set:
  32 such stays, each costing a call plus retries, every run, indefinitely,
  against an API whose rate limit is still unknown (U15). A new
  collection_attempt table records what each attempt did so the collector can
  back off — three free attempts, then 1h/2h/4h… capped at a week. Any success
  resets the counter, so a stay that reopens is picked straight back up. It is
  not a fact table: nothing in it reaches a baseline or a score.

Also moves the grid query out of scripts/ into packages/data, where the
project's SQL belongs.

Verified by a cold-start rehearsal: day 1 collects 210 stays into 2,580 rates
and 133 room types; day 2 correctly re-requests only the 32 non-yielding stays
and skips everything still inside its tier interval; by run 4 those are backed
off and the run is a clean no-op.

Two things recorded rather than fixed, because they are the source's behaviour
and not ours: status 200 with `{"count": 0, "filtered": 65}` means the API
filtered every room away server-side rather than the hotel being empty, and a
daily cron deliberately under-serves the HOT tier, which wants 6 hours.

270 tests, 9 schema checks.

+++ packages/ingest/src/scheduler/tiers.ts
@@ -100,7 +100,9 @@ export async function planCollection(
          JOIN hotel h ON h.id = o.hotel_id
         WHERE h.is_active AND h.collection_tier <> 'OFF'
           AND o.check_in >= CURRENT_DATE
-          AND o.check_in <= CURRENT_DATE + $1
+          -- ::int is required. Bare $1 leaves Postgres unable to choose
+          -- between date+integer and date+interval: "operator is not unique".
+          AND o.check_in <= CURRENT_DATE + $1::int
         GROUP BY o.hotel_id, h.wah_hotel_id, o.check_in, o.nights, o.adults
      ),
      viewed AS (
